# src/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a simple bookstore server with a book inventory, users and\
    \ a shopping cart."
  version: "1.0.0"
  title: "Shelf Awareness Bookstore"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8090"
basePath: "/v2"
tags:
  - name: "book"
    description: "Book's fact sheet"
  - name: "author"
    description: "Information about author"
schemes:
  - "https"
paths:
  /books:
    get:
      tags:
        - "book"
      summary: "Books available in the inventory"
      description: "List of books available in the inventory"
      operationId: "booksGET"
      produces:
        - "application/json"
      parameters:
        - name: "offset"
          in: "query"
          description: "Pagination offset. Default is 0."
          required: false
          type: "integer"
        - name: "limit"
          in: "query"
          description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
          required: false
          type: "integer"
          default: 20
      responses:
        200:
          description: "A collection of Books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Book"
  /books/{bookId}:
    get:
      tags:
        - "book"
      summary: "Find book by ID"
      description: "Returns a book"
      operationId: "getBookById"
      produces:
        - "application/json"
      parameters:
        - name: "bookId"
          in: "path"
          description: "ID of book to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Book not found"
      x-swagger-router-controller: "Book"
  /authors:
    get:
      tags:
        - "author"
      summary: "Authors available in the inventory"
      description: "List of authors available in the inventory"
      operationId: "authorsGET"
      produces:
        - "application/json"
      parameters:
        - name: "offset"
          in: "query"
          description: "Pagination offset. Default is 0."
          required: false
          type: "integer"
        - name: "limit"
          in: "query"
          description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
          required: false
          type: "integer"
          default: 20
      responses:
        200:
          description: "A collection of Authors"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Author"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Author"
  /authors/{authorId}:
    get:
      tags:
        - "author"
      summary: "Find author by ID"
      description: "Returns a author"
      operationId: "getAuthorById"
      produces:
        - "application/json"
      parameters:
        - name: "authorId"
          in: "path"
          description: "ID of author to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Author"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Author not found"
      x-swagger-router-controller: "Author"
  /books/byAuthor/{authorId}:
    get:
      tags:
        - "book"
      summary: "Books by author available in the inventory"
      description: "List of books by author available in the inventory"
      operationId: "booksByAuthorGET"
      produces:
        - "application/json"
      parameters:
        - name: "authorId"
          in: "path"
          description: "ID of book of reviews"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "A collection of books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Book"
  /users:
    get:
      tags:
        - "user"
      summary: "Active users"
      description: "List of users available in the inventory"
      operationId: "usersGET"
      produces:
         - "application/json"
      parameters:
         - name: "offset"
           in: "query"
           description: "Pagination offset. Default is 0."
           required: false
           type: "integer"
         - name: "limit"
           in: "query"
           description: "Maximum number of items per page. Default is 20 and cannot exceed\
               \ 500."
           required: false
           type: "integer"
           default: 20
      responses:
        200:
          description: "A collection of Users"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "User"
  /users/loggedInUser:
    get:
      tags:
        - "user"
      summary: "Find user that is currently logged in"
      description: "Returns a user"
      operationId: "loggedInUserGET"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        404:
          description: "User not found"
      x-swagger-router-controller: "User"
  /users/logoutUser:
    get:
      tags:
        - "user"
      summary: "Log out the user that is logged in"
      description: "returns the result of logging out"
      operationId: "logoutUserGET"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        404:
          description: "User not found"
      x-swagger-router-controller: "User"
  /users/{userId}:
    get:
      tags:
        - "user"
      summary: "Find user by ID"
      description: "Returns a user"
      operationId: "getUserById"
      produces:
        - "application/json"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of user to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "User"
  /reviews:
    get:
      tags:
        - "review"
      summary: "Reviews available in the inventory"
      description: "List of reviews available in the inventory"
      operationId: "reviewsGET"
      produces:
        - "application/json"
      parameters:
        - name: "offset"
          in: "query"
          description: "Pagination offset. Default is 0."
          required: false
          type: "integer"
        - name: "limit"
          in: "query"
          description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
          required: false
          type: "integer"
          default: 20
      responses:
        200:
          description: "A collection of reviews"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Review"
  /reviews/{reviewId}:
    get:
      tags:
        - "review"
      summary: "Find review by ID"
      description: "Returns a review"
      operationId: "getReviewById"
      produces:
        - "application/json"
      parameters:
        - name: "reviewId"
          in: "path"
          description: "ID of review to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Review"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Review not found"
      x-swagger-router-controller: "Review"
  /reviews/byBook/{bookId}:
    get:
      tags:
        - "review"
      summary: "Find reviews by book ID"
      description: "Returns reviews for one book"
      operationId: "getReviewsByBook"
      produces:
        - "application/json"
      parameters:
        - name: "bookId"
          in: "path"
          description: "ID of book of reviews"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "A collection of reviews"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Reviews not found"
      x-swagger-router-controller: "Review"
  /events:
    get:
      tags:
        - "event"
      summary: "Events available in the inventory"
      description: "List of events available in the inventory"
      operationId: "eventsGET"
      produces:
        - "application/json"
      parameters:
        - name: "offset"
          in: "query"
          description: "Pagination offset. Default is 0."
          required: false
          type: "integer"
        - name: "limit"
          in: "query"
          description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
          required: false
          type: "integer"
          default: 20
      responses:
        200:
          description: "A collection of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Event"
  /events/{eventId}:
    get:
      tags:
        - "event"
      summary: "Find event by ID"
      description: "Returns a event"
      operationId: "getEventById"
      produces:
        - "application/json"
      parameters:
        - name: "eventId"
          in: "path"
          description: "ID of event to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Review not found"
      x-swagger-router-controller: "Event"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfull login"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "userRegisterPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
      - name: "email"
        in: "formData"
        required: true
        type: "string"
      - name: "address"
        in: "formData"
        required: true
        type: "string"
      - name: "creditcard"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfull registration"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /reviews/create:
    post:
      tags:
      - "review"
      summary: "Create review"
      description: "Create new review"
      operationId: "createReviewPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "formData"
        required: true
        type: "integer"
        format: "int64"
      - name: "userId"
        in: "formData"
        required: true
        type: "integer"
        format: "int64"
      - name: "comment"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfull creation"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Review"
  /cart/{cartId}:
    get:
      tags:
      - "cart"
      summary: "View the content of the cart"
      operationId: "cartCartIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "cartId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "succesful operartion"
          schema:
            $ref: "#/definitions/Cart"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Cart"
definitions:
  User:
    type: "object"
    required:
      - "email"
      - "password"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
      address:
        type: "string"
      creditcard:
        type: "string"
    title: "User"
    description: "A user"
    example:
      id: 1
      name: "Andrea Bertini"
      email: "andrea.bertini@mail.polimi.it"
      password: 123456
      address: "DEIB"
      creditcard: "xyzabc"

  Book:
    type: "object"
    required:
      - "author"
      - "price"
      - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      author:
        $ref: "#/definitions/Author"
      cover:
        type: "string"
      abstract:
        type: "string"
      genre:
        type: "string"
        description: "book genre"
        enum: ["Adventure", "comic", "Crime", "Erotic", "Fiction", "Fantasy", "Historical", "Horror", "Magic", "Realism", "Mystery", "Paranoid", "Philosophical", "Political", "Romance", "Saga", "Satire", "Science", "Thriller", "Urban", "Western"]
      price:
        $ref: "#/definitions/Amount"
      status:
        type: "string"
        description: "book availability in the inventory"
        enum: ["available", "out of stock"]
    title: "Book"
    description: "A book for sale in the store"
    example:
      id: 0
      title: "Il deserto dei tartari"
      author:
        id: 0
        name: "Dino Buzzati"
        picture:
        bio: "Nasce il 16 ottobre 1906 a San Pellegrino, nei pressi di Belluno. Sin dalla giovinezza si manifestano in lui gli interessi, i temi e le passioni del futuro scrittore, ai quali resterà fedele per tutta la vita la poesia, la musica (studia violino e pianoforte e non bisogna dimenticare che in futuro scriverà anche alcuni libretti d'Opera)"
      cover:
      abstract: "La trama segue la vita del sottotenente Giovanni Drogo dal momento in cui, divenuto ufficiale, viene assegnato come prima nomina alla Fortezza Bastiani, molto distante dalla città"
      genre: "Fiction"
      price:
        value: 10
        currency: "eur"
      status: "available"

  Author:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      bio:
        type: "string"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Author"
    description: "An author"
    example:
      id: 0
      name: "Dino Buzzati"
      bio: "Nasce il 16 ottobre 1906 a San Pellegrino, nei pressi di Belluno. Sin dalla giovinezza si manifestano in lui gli interessi, i temi e le passioni del futuro scrittore, ai quali resterà fedele per tutta la vita la poesia, la musica (studia violino e pianoforte e non bisogna dimenticare che in futuro scriverà anche alcuni libretti d'Opera)"
      picture:
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"

  Cart:
    type: "object"
    properties:
      total:
        $ref: "#/definitions/Amount"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Cart"
    description: "Order for books"
    example:
      total:
        currency: "eur"
        value: 80082819046101.171875
      books:
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"

  Amount:
    type: "object"
    required:
      - "currency"
      - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000000000000000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price"
    example:
      currency: "eur"
      value: 602745618307040.375

  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "Money system"
    example: "eur"

  Review:
    type: object
    required:
      - comment
      - user
      - book
    properties:
        id:
          type: integer
          format: int64
        comment:
          type: string
        user:
          $ref: "#/definitions/User"
        book:
          $ref: "#/definitions/Book"
    title: Review
    description: A review
    example:
      id: 1
      comment: Love it!
      user:
        id: 1
        name: Andrea Bertini
        email: andrea.bertini@mail.polimi.it
        password: 123456
        address: DEIB
        creditcard: xyzabc
      book:
        id: 1
        title: Il deserto dei tartari
        author:
          id: 1
          name: Dino Buzzati
          biography: Nasce il 16 ottobre 1906 a San Pellegrino, nei pressi di Belluno. Sin dalla giovinezza si manifestano in lui gli interessi, i temi e le passioni del futuro scrittore, ai quali resterà fedele per tutta la vita la poesia, la musica (studia violino e pianoforte e non bisogna dimenticare che in futuro scriverà anche alcuni libretti d'Opera), il disegno, e la montagna, vera compagna dell'infanzia, a cui è anche dedicato il suo primo romanzo, "Barnabo delle montagne".

  Event:
    type: object
    required:
      - location
      - held_on
      - book
    properties:
        id:
          type: integer
          format: int64
        location:
          type: string
        held_on:
          type: string
          format: date
        book:
          $ref: "#/definitions/Book"
    title: Event
    description: An event
    example:
      id: 1
      location: Biblioteca Leonardo
      held_on:
      book:
        id: 1
        title: Il deserto dei tartari
        author:
          id: 1
          name: Dino Buzzati
          biography: Nasce il 16 ottobre 1906 a San Pellegrino, nei pressi di Belluno. Sin dalla giovinezza si manifestano in lui gli interessi, i temi e le passioni del futuro scrittore, ai quali resterà fedele per tutta la vita la poesia, la musica (studia violino e pianoforte e non bisogna dimenticare che in futuro scriverà anche alcuni libretti d'Opera), il disegno, e la montagna, vera compagna dell'infanzia, a cui è anche dedicato il suo primo romanzo, "Barnabo delle montagne".

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"